% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brain_gen.R
\name{brain_infuse}
\alias{brain_infuse}
\title{Create gif from brain image data}
\usage{
brain_infuse(
  file,
  color_palette = RColorBrewer::brewer.pal(name = "Set3", n = 8),
  color_background = "#000000",
  infuse_rate = 0.9,
  loop = TRUE,
  delay = 0.05,
  trim_start = 0.25,
  trim_end = 0.05,
  output_dir = "brain_infuse",
  gif_file = "brain_infuse.gif",
  clean_files = TRUE
)
}
\arguments{
\item{file}{Character; File path of extension \code{.nii.gz}}

\item{color_palette}{Character vector of color codes}

\item{color_background}{Character hex code giving color background}

\item{infuse_rate}{Numeric; Between 0.5-0.99 e.g., \code{infusion_rate} must be >= 50\%.
Higher values infuse color at greater intervals.}

\item{loop}{Logical; If \code{TRUE} images will play then reverse}

\item{delay}{Numeric; Time delay for each image in seconds. Increase to slow gif}

\item{trim_start}{Numeric; Increasing value will remove frames at beginning of
gif e.g., empty frames with only background color}

\item{output_dir}{Character; If missing, folder will be created in current working directory.}

\item{gif_file}{Character; File name of gif}

\item{clean_files}{Logical; Set to \code{TRUE} to remove png files in \code{output_dir}}
}
\value{

}
\description{
Create gif from brain image data
}
\examples{
library(neurogen)
nii_path <- system.file(package = "neurogen", "data", "Ty_brain.nii.gz")
brain_infuse(file = nii_path,
             color_palette = RColorBrewer::brewer.pal(name = "Set3", n = 8),
             infuse_rate = 0.1,
             trim_start = 0.31,
             trim_end = 0.08)
             )

}
