\name{cancer}
\alias{cancer}
\docType{data}
\title{Datos de cancer de piel}
\description{Datos de cancer de piel.}
\usage{data(cancer)}
\format{
  Una arreglo numerico con 50 observaciones 3 variables explicativas y 1 variable respuesta.
  \describe{
    \item{Columna1}{Tipo de piel: 1 = clara, 2 = medio y 3 = oscura.}
    \item{Columna2}{Examen Fisico: Valor entre 0 y 1 que representa un valor de probabilidad de que el paciente tenga cancer de piel. Un examen fisico asigna el valor a cada paciente.}    
    \item{Columna3}{Prueba de Sangre: Valores entre 0 y 10 que representa una prueba de sangre en la cual se observa la presencia de celulas cancerosas.}    
    \item{Columna4}{Biopsia: El diagnostico maligno = 1 benigno = 0.}               
    }
}
\source{
  Neuro Shell Series 
}
\keyword{datasets}
