% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brain_gen.R
\name{brain_warp}
\alias{brain_warp}
\title{Title}
\usage{
brain_warp(
  file,
  color_palette = hcl.colors(n = 6, palette = "Sunset"),
  loop = TRUE,
  delay = 0.05,
  width = 600,
  height = 800,
  z_offset = 0.25,
  output_dir = "brain_warp",
  gif_file = "brain_warp.gif",
  clean_files = TRUE
)
}
\arguments{
\item{delay}{Numeric; time to show each image in seconds}

\item{height}{}
}
\value{

}
\description{
Title
}
\examples{

}
